Expanded time and interrupts

+++ arch/include/synos/arch/io.h
@@ -3,8 +3,8 @@
 #include <inttypes.h>
 #include <stdbool.h>
 
-uint8_t inb(uint32_t);
-void outb(uint32_t, uint8_t);
+uint8_t inb(uint16_t);
+void outb(uint16_t, uint8_t);
 void io_wait();
 
 #endif
